Ohai darwin hardware plugin should catch command timeouts.
On Big Sur, the ohai darwin hardwqare plugin can timeout when running
`system_profiler SPHardwareDataType -xml`. (On my laptop, it takes 6
minutes for this command to complete and return no data.) When this
command times out the plugin aborts and doesn't return any hardware
data. When this command times out the plugin continue so that
in can return some data.

Signed-off-by: Edward Pilatowicz epilatow@fb.com